+++ piglegsurgeryweb/uploader/urls.py
@@ -11,6 +11,7 @@ urlpatterns = [
     path("upload/", views.model_form_upload, name="model_form_upload"),
     path("thanks/", views.thanks, name="thanks"),
     path("message/", views.message, name="message"),
+    path("message/<str:headline>/<str:text>/<str:next>/<str:next_text>", views.message, name="message"),
     path(
         "<int:filename_id>/run/", views.run, name="run"
     ),  # used for debugging purposes
@@ -26,6 +27,7 @@ urlpatterns = [
         name="swap_is_microsurgery",
     ),
     path("web_report/<str:filename_hash>/", views.web_report, name="web_report"),
+    path("web_report/<str:filename_hash>/<str:review_edit_hash>/", views.web_report, name="web_report"),
     path(
         "owners_reports/<str:owner_hash>/",
         views.owners_reports_list,
@@ -48,6 +50,8 @@ urlpatterns = [
     # path("about_ev_cs/", views.about_ev_cs, name="about_ev_cs"),
     path("test/", views.test, name="test"),
     path("show_logs/<str:filename_hash>", views.show_logs, name="show_logs"),
-    path("login/", auth_views.LoginView.as_view(), name="login")
+    path("login/", auth_views.LoginView.as_view(), name="login"),
+    path("go_to_video_for_annotation/", views.go_to_video_for_annotation, name="go_to_video_for_annotation_random"),
+    path("go_to_video_for_annotation/<str:email>/", views.go_to_video_for_annotation, name="go_to_video_for_annotation_email"),
     # path("about_ev_en/", views.about_ev_en, name="about_ev_en"),
 ]